2016/02/27

+++ Application/Admin/Controller/SystemController.class.php
@@ -23,31 +23,33 @@ class SystemController extends CommonController {
 	 * @author lhk(2016/01/28)
 	 */
 	public function logData() {
-		$conditions = array (
-				'start_date' => I ( 'get.start_date' ),
-				'end_date' => I ( 'get.end_date' ),
-				'keywords' => I ( 'get.keywords' ) 
-		);
-		$adminLogModel = D ( 'AdminLog' );
-		$list = $adminLogModel->logData ( $conditions );
-		
-		$data = array ();
-		foreach ( $list as $k => $v ) {
-			$data [$k] = array (
-					'checkbox' => "<input value=\"{$v['admin_log_id']}\" name=\"id\" type=\"checkbox\" />",
-					'admin_log_id' => $v ['admin_log_id'],
-					'admin_realname' => $v ['admin_realname'],
-					'admin_log_type' => $v ['admin_log_type'],
-					'admin_log_content' => $v ['admin_log_content'],
-					'admin_log_time' => date ( 'Y-m-d H:i:s', $v ['admin_log_time'] ),
-					'admin_log_ip' => $v ['admin_log_ip'],
-					'handle' => "<a title=\"删除\" href=\"javascript:;\" onclick=\"delete_one('{$v['admin_log_id']}')\" class=\"ml-5\" style=\"text-decoration:none\"><i class=\"Hui-iconfont\">&#xe6e2;</i></a>" 
+		if (IS_AJAX) {
+			$conditions = array (
+					'start_date' => I ( 'get.start_date' ),
+					'end_date' => I ( 'get.end_date' ),
+					'keywords' => I ( 'get.keywords' ) 
 			);
+			$adminLogModel = D ( 'AdminLog' );
+			$list = $adminLogModel->logData ( $conditions );
+			
+			$data = array ();
+			foreach ( $list as $k => $v ) {
+				$data [$k] = array (
+						'checkbox' => "<input value=\"{$v['admin_log_id']}\" name=\"id\" type=\"checkbox\" />",
+						'admin_log_id' => $v ['admin_log_id'],
+						'admin_realname' => $v ['admin_realname'],
+						'admin_log_type' => $v ['admin_log_type'],
+						'admin_log_content' => $v ['admin_log_content'],
+						'admin_log_time' => date ( 'Y-m-d H:i:s', $v ['admin_log_time'] ),
+						'admin_log_ip' => $v ['admin_log_ip'],
+						'handle' => "<a title=\"删除\" href=\"javascript:;\" onclick=\"delete_one('{$v['admin_log_id']}')\" class=\"ml-5\" style=\"text-decoration:none\"><i class=\"Hui-iconfont\">&#xe6e2;</i></a>" 
+				);
+			}
+			
+			$this->ajaxReturn ( array (
+					'data' => $data 
+			) );
 		}
-		
-		$this->ajaxReturn ( array (
-				'data' => $data 
-		) );
 	}
 	
 	/**
